+++ src/ThumbCreator.php
@@ -1,5 +1,6 @@
 <?php
 declare(strict_types=1);
+
 /**
  * This file is part of php-thumber.
  *
@@ -23,6 +24,7 @@ use Symfony\Component\Filesystem\Filesystem;
 use Thumber\Exception\NotReadableImageException;
 use Thumber\Exception\UnsupportedImageTypeException;
 use Tools\Exception\NotWritableException;
+use Tools\Exceptionist;
 
 /**
  * Utility to create a thumb.
@@ -72,6 +74,8 @@ class ThumbCreator
      * It sets the file path and extension.
      * @param string $path Path of the image from which to create the
      *  thumbnail. It can be a full path or a remote url
+     * @throws \Tools\Exception\FileNotExistsException
+     * @throws \Tools\Exception\NotReadableException
      * @uses $ImageManager
      * @uses $arguments
      * @uses $path
@@ -79,7 +83,7 @@ class ThumbCreator
     public function __construct(string $path)
     {
         if (!is_url($path)) {
-            is_readable_or_fail($path);
+            Exceptionist::isReadable($path);
         }
         $this->ImageManager = new ImageManager(['driver' => THUMBER_DRIVER]);
         $this->arguments[] = $this->path = $path;
@@ -268,17 +272,15 @@ class ThumbCreator
      */
     public function save(array $options = []): string
     {
-        if (!$this->callbacks) {
-            throw new BadMethodCallException(sprintf('No valid method called before the `save()` method'));
-        }
+        Exceptionist::isTrue($this->callbacks, 'No valid method called before the `save()` method', BadMethodCallException::class);
 
         $options = $this->getDefaultSaveOptions($options);
         $target = $options['target'];
-        $options['format'] = $target ? $this->getDefaultSaveOptions([], $target)['format'] : $options['format'];
+        $format = $target ? $this->getDefaultSaveOptions([], $target)['format'] : $options['format'];
 
         if (!$target) {
-            $this->arguments[] = [THUMBER_DRIVER, $options['format'], $options['quality']];
-            $target = sprintf('%s_%s.%s', md5($this->path), md5(serialize($this->arguments)), $options['format']);
+            $this->arguments[] = [THUMBER_DRIVER, $format, $options['quality']];
+            $target = sprintf('%s_%s.%s', md5($this->path), md5(serialize($this->arguments)), $format);
         }
 
         $target = (new Filesystem())->isAbsolutePath($target) ? $target : add_slash_term(THUMBER_TARGET) . $target;
@@ -292,7 +294,7 @@ class ThumbCreator
                 call_user_func($callback, $imageInstance);
             }
 
-            $content = $imageInstance->encode($options['format'], $options['quality']);
+            $content = $imageInstance->encode($format, $options['quality']);
             $imageInstance->destroy();
             try {
                 (new Filesystem())->dumpFile($target, $content);