+++ tests/TestCase/ThumbManagerTest.php
@@ -1,5 +1,6 @@
 <?php
 declare(strict_types=1);
+
 /**
  * This file is part of php-thumber.
  *
@@ -86,7 +87,7 @@ class ThumbManagerTest extends TestCase
 
         //With a no existing file
         $this->expectException(NotReadableException::class);
-        $this->expectExceptionMessage('File or directory `noExisting` is not readable');
+        $this->expectExceptionMessage('File or directory `noExisting` does not exist');
         $this->ThumbManager->get('noExisting');
     }